Move sendFulfill
